Create ListView to add/modify/view/delete flashcards
Added new view - ListView - to view all
flashcards in the current database, as
well as interact with the cards to update
or delete them. Furthermore, changed the
flashcard model to include attempts, an
element that could be used to push back
the cards a specific amount in later
stages of development.